Slightly change API
